+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Version 1.2.0 *(unreleased)*
 * New args can be added for npmSetup and yarnSetup (#226)
 * Uses --no-save for npmSetup and yarnSetup (#222)
 * Added execOverrides and ignoreExitValue on npmSetupTask (#196)
+* Added gruntFile in grunt extension to use different file (#189)
 * ...
 
 Version 1.1.1 *(2017-01-16)*